+++ frontend/src/drivers/supaplex/internal.ts
@@ -61,11 +61,13 @@ export interface ILevelFooter extends IWithDemo, IWithSignature {
   readonly usePlasmaTime: number | undefined;
   readonly useZonker: boolean;
   readonly useSerialPorts: boolean;
+  readonly useInfotronsNeeded: number | undefined;
   setUsePlasma(on: boolean): this;
   setUsePlasmaLimit(n: number | undefined): this;
   setUsePlasmaTime(n: number | undefined): this;
   setUseZonker(on: boolean): this;
   setUseSerialPorts(on: boolean): this;
+  setUseInfotronsNeeded(n: number | undefined): this;
 }
 
 export const SPEC_PORT_MAX_COUNT = 10;
@@ -76,4 +78,5 @@ export const enum LocalOpt {
   UsePlasmaTime = "usePlasmaTime",
   UseZonker = "useZonker",
   UseSerialPorts = "useSerialPorts",
+  UseInfotronsNeeded = "useInfotronsNeeded",
 }