+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.15.0 (DEV)
 
+- Add: Option to SO Test Page to override number of infotrons needed.
 - Fix: Produced MPX files now should work better in Megaplex. Dropped support of
   incorrect MPX header (with spaces) when saving MPX file with 1 level.